+++ docs/CHANGELOG.rst
@@ -18,6 +18,8 @@ Changed
 Fixed
 -----
 - Fix typo in ``scheduling_class`` variable in several Python processes
+- Handle cases of improper tags passed to ``read_group`` argument of
+  the ``bqsr`` process
 
 
 ===================